+++ tigon-flow/src/main/java/co/cask/tigon/flow/DeployClient.java
@@ -118,7 +118,7 @@ public class DeployClient {
     return deployFlow(clz);
   }
 
-  Location deployFlow(Class<?> flowClz, File... bundleEmbeddedJars)
+  public Location deployFlow(Class<?> flowClz, File... bundleEmbeddedJars)
     throws Exception {
     Preconditions.checkNotNull(flowClz, "Flow cannot be null.");
     Location deployedJar = locationFactory.create(createDeploymentJar(
