--- tigon-flow/src/main/resources/mapred-site-local.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0"?>
-<?xml-stylesheet type="text/xsl" href="configuration.xsl"?>
-
-<!--
-  Copyright 2014 Cask Data, Inc.
-
-  Licensed under the Apache License, Version 2.0 (the "License"); you may not
-  use this file except in compliance with the License. You may obtain a copy of
-  the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-  WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-  License for the specific language governing permissions and limitations under
-  the License.
-  -->
-
-<configuration>
-  <property>
-    <name>mapreduce.framework.name</name>
-    <value>local</value>
-  </property>
-  <property>
-    <name>mapreduce.jobtracker.address</name>
-    <value>clocal</value>
-    <description>
-      See LocalClientProtocolProvider for more info
-    </description>
-  </property>
-  <property>
-    <name>mapreduce.job.counters.counter.name.max</name>
-    <value>256</value>
-  </property>
-  <property>
-    <name>mapreduce.local.map.tasks.maximum</name>
-    <value>8</value>
-    <description>The maximum number of map tasks to run in parallel in LocalJobRunner</description>
-  </property>
-  <property>
-    <name>mapreduce.local.reduce.tasks.maximum</name>
-    <value>8</value>
-    <description>The maximum number of reduce tasks to run in parallel in LocalJobRunner</description>
-  </property>
-</configuration>
